Backend/requirements.txt
sqlalchemy
sqlcipher3
yfinance
pydantic
pydantic-settings
pandas
python-dotenv
httpx
bcrypt
PyJWT
requests
requests-cache
requests-ratelimiter
pyrate-limiter
pytest
fastapi
uvicorn
python-multipart
nsetools
beautifulsoup4
lxml
matplotlib
aiosmtpd
